+++ src/main/java/com/navnath/config/SwaggerConfig.java
@@ -1,0 +1,5 @@
+package com.navnath.config;
+
+public class SwaggerConfig {
+
+}
