+++ src/main/java/com/navnath/controller/BankAccountController.java
@@ -1,7 +1,10 @@
 package com.navnath.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.navnath.service.BankAccountServices;
@@ -10,12 +13,13 @@ import com.navnath.service.BankAccountServices;
 public class BankAccountController {
 	
 	@Autowired
-	BankAccountServices bankAccountServices;
+	private  BankAccountServices bankAccountServices;
 	
 	
 	 @GetMapping(value = "/testConnectivity")
-	public void testConnectivity() {
-		bankAccountServices.testConnectivity();
+	public ResponseEntity<String> testConnectivity() {
+		return ResponseEntity.ok("{Status }" + bankAccountServices.testConnectivity() );
+			
 	}
