--- src/main/java/com/navnath/schema/Contact.java
@@ -1,21 +1,0 @@
-package com.navnath.schema;
-
-public class Contact {
-
-	private int phone;
-	private String email;
-	public int getPhone() {
-		return phone;
-	}
-	public void setPhone(int phone) {
-		this.phone = phone;
-	}
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	
-	
-}
