--- src/main/java/com/navnath/schema/Identification.java
@@ -1,21 +1,0 @@
-package com.navnath.schema;
-
-public class Identification {
-
-	private int drivingLicenceNumber;
-	private String drivingLicenceState;
-	public int getDrivingLicenceNumber() {
-		return drivingLicenceNumber;
-	}
-	public void setDrivingLicenceNumber(int drivingLicenceNumber) {
-		this.drivingLicenceNumber = drivingLicenceNumber;
-	}
-	public String getDrivingLicenceState() {
-		return drivingLicenceState;
-	}
-	public void setDrivingLicenceState(String drivingLicenceState) {
-		this.drivingLicenceState = drivingLicenceState;
-	}
-	
-	
-}
