--- src/main/java/com/navnath/schema/PostalAddress.java
@@ -1,36 +1,0 @@
-package com.navnath.schema;
-
-public class PostalAddress {
-
-	private String address;
-	private int postcode;
-	private String state;
-	private String country;
-	public String getAddress() {
-		return address;
-	}
-	public void setAddress(String address) {
-		this.address = address;
-	}
-	public int getPostcode() {
-		return postcode;
-	}
-	public void setPostcode(int postcode) {
-		this.postcode = postcode;
-	}
-	public String getState() {
-		return state;
-	}
-	public void setState(String state) {
-		this.state = state;
-	}
-	public String getCountry() {
-		return country;
-	}
-	public void setCountry(String country) {
-		this.country = country;
-	}
-	
-	
-	
-}
