+++ src/main/java/com/navnath/service/BankAccountServices.java
@@ -1,6 +1,9 @@
 package com.navnath.service;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public interface BankAccountServices {
 
-	void testConnectivity();
+	String testConnectivity();
 }
