+++ src/main/resources/application.properties
@@ -1,1 +1,9 @@
-server.port=8081
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+#spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.h2.console.enabled =true
+server.port = 8080
+
+#http://localhost:8080/h2-console
